requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
requests==2.31.0
motor==3.3.2
dnspython==2.4.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
fastapi-mail==1.4.1
google-auth==2.27.0
gunicorn==21.2.0
